requirements.txt
streamlit
pandas
python-dotenv
fuzzywuzzy
python-Levenshtein  # Optional, but makes fuzzywuzzy faster
requests 